app/requirements.txt
ultralytics
gradio
torch
Pillow
